Refator prebuilt Agora functions for joining streams.
